Combine partial events using the `last` flag supplied to WebSocketListener
...instead of passing data through a PipedWriter/PipedReader pair and parsing it using GSON's streaming mode.

A significant simplification to Helog's internal architecture.

With this change, `--raw` output writes each JSON object on a new line.

Also, Helog will now tolerate malformed JSON - it will write an error message and the malformed payload to stderr, and then continue.

Fixes #1, as Helog now tolerates the WebSocket event thread changing, but it's possible that Helog will appear to hang in the (currently unknown) situation that causes #1.